backend/pyproject.toml
[tool.poetry]
name = "be-todo-list"
version = "0.1.0"
description = ""
authors = ["nugrahari <harin260@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
